Use inspect.unwrap, rename _support -> decorator_mod
