+++ Lib/test/test_doctest/doctest_lineno.py
@@ -70,7 +70,7 @@ str_wrapper = object().__str__
 
 
 # https://github.com/python/cpython/issues/115392
-from test.test_doctest._support import decorator
+from test.test_doctest.decorator_mod import decorator
 
 @decorator
 @decorator
